+++ src/components/burger-ingredients/burger-ingredients.js
@@ -45,16 +45,15 @@ const BurgerIngredients = (props) => {
   <nav className={BurgerIngredientsStyles.ingredients}>
     
         <nav className={BurgerIngredientsStyles.buns}>
-           <MenuItem id='bun' text='Булочки' function={buns.map(element => <Ingredient key={element._id} data={element}/>)} />
+           <MenuItem value='bun' text='Булочки' function={buns.map(element => <Ingredient id={element._id} data={element}/>)} />
         </nav>
         <nav className={BurgerIngredientsStyles.sauses}>
-            <MenuItem id='sauce' text='Соусы' function={sauces.map(element => <Ingredient key={element._id} data={element}/>)} />
+            <MenuItem value='sauce' text='Соусы' function={sauces.map(element => <Ingredient id={element._id} data={element}/>)} />
         </nav>
         <nav className={BurgerIngredientsStyles.mains}>
-            <MenuItem id='main' text='Начинки' function={mains.map(element => <Ingredient key={element._id} data={element}/>)} />          
+            <MenuItem value='main' text='Начинки' function={mains.map(element => <Ingredient id={element._id} data={element}/>)} />          
         </nav>
   </nav>
-       
       </section>
     )
   }
@@ -66,4 +65,5 @@ const BurgerIngredients = (props) => {
       function:PropTypes.func
   };
   
-export default BurgerIngredients; 
+export default BurgerIngredients;
+
