+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -1,139 +1,123 @@
 .ingredientsMenu {
     /* BurgerIngredients */
-     position: absolute;
+    display: flex;
+    flex-direction: column;
      width: 600px;
      height: 912px;
      left: 380px;
      top: 128px;
+
    
      /* interface/bg */
    
      background: #131316;
-     /* interface/accent */
+
+}
    
-     border: 2px solid #4C4CFF;
-   }
-   
-   .tabMenu {
-     /* Auto layout */
-   
+.tabMenu { 
    display: flex;
    flex-direction: row;
    align-items: flex-start;
    padding: 0px;
    
-   position: absolute;
+
    width: 600px;
    height: 56px;
    left: 0px;
    top: 100px;
-   
-   
-   }
+}
    
    
    
-   .description {
+.description {   
+display: flex;
+flex-wrap: wrap;
    
-     display: flex;
-     flex-wrap: wrap;
-   
-     padding: 0;
+ padding: 0;
    
    
      
-     height: 48px;
-     left: 0%;
-     right: 0%;
-     bottom: 0px;
+ height: 48px;
+ left: 0%;
+  right: 0%;
+ bottom: 0px;
    
-     /* desktop/text */
+ /* desktop/text */
    
-     font-family: 'JetBrains Mono';
-     font-style: normal;
-     font-weight: 400;
-     font-size: 16px;
+  font-family: 'JetBrains Mono';
+  font-style: normal;
+  font-weight: 400;
+  font-size: 16px;
    
-     /* or 150% */
+  /* or 150% */
    
-     text-align: center;
-   
-     /* text/primary-light (standart text) */
-   
-     color: #F2F2F3;
-   }
+  text-align: center;
+}
     
-   .ingredients {
-     position: absolute;
-     width: 600px;
-     height: 756px;
-     left: 0px;
-     top: 156px;
-     overflow-y: scroll;
+.ingredients {
+
+  width: 600px;
+  height: 756px;
+  left: 0px;
+  top: 156px;
+  overflow-y: scroll;
+  overflow-x: hidden;
    
-     border-radius: 0px;
-   }
+  border-radius: 0px;
+}
    
-   .buns{
+.buns{
+
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: flex-start;
+  padding: 0px 16px;
+  gap: 24px;
+
+  width: 600px;
+  height: 208px;
+  left: 0px;
+  top: 16px;
+}
+   
+.sauses {
      /* Auto layout */
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: flex-start;
+  padding: 80px 16px;
+  gap: 24px;
    
-     display: flex;
-     flex-direction: row;
-     justify-content: center;
-     align-items: flex-start;
-     padding: 0px 16px;
-     gap: 24px;
+
+  width: 600px;
+  height: 472px;
+  left: 0px;
+  top: 396px;
+}
    
-     position: absolute;
-     width: 600px;
-     height: 208px;
-     left: 0px;
-     top: 16px;
-   }
+.mains {
+
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: flex-start;
+  padding: 0px 16px;
+  gap: 24px;
    
-   .sauses {
-     /* Auto layout */
-     display: flex;
-     flex-direction: row;
-     justify-content: center;
-     align-items: flex-start;
-     padding: 0px 16px;
-     gap: 24px;
+
+  width: 600px;
+  height: 1184px;
+  left: 0px;
+  top: 962px;
+}
    
-     position: absolute;
-     width: 600px;
-     height: 472px;
-     left: 0px;
-     top: 396px;
-   }
+.headline {
    
-   .mains {
-     /* Auto layout */
-   
-   display: flex;
-   flex-direction: row;
-   justify-content: center;
-   align-items: flex-start;
-   padding: 0px 16px;
-   gap: 24px;
-   
-   position: absolute;
-   width: 600px;
-   height: 1184px;
-   left: 0px;
-   top: 962px;
-   }
-   
-   .headline {
-   
-     width: 600px;
-     height: 30px;
-     left: 0px;
-     
-     /* interface/bg */
-     
-     background: #131316;
-     /* interface/accent */
-     
-     border: 1px dashed #4C4CFF;
-   }
+  width: 600px;
+  height: 30px;
+  left: 0px;
+
+  background: #131316;
+}
