+++ src/components/burger-ingredients/ingredient/ingredient.js
@@ -6,23 +6,12 @@ import { CurrencyIcon } from '@ya.praktikum/react-developer-burger-ui-components
 
 const Ingredient = (props) => {
 
-    class CurrencyIcons extends React.Component{
-        render(){
-            return(
-                <ul className={IngredientsStyles.currency} >
-                    {this.props.icon}
-                </ul>
-            )
-        }
-    }
-    
-
     return (
-      <li className={IngredientsStyles.ingredient} key={props.data._id} >
+      <li className={IngredientsStyles.ingredient} id={props.data._id} >
         <img className="ml-4 mr-4" src={props.data.image} alt={props.data.name}/>
         <div className={`${IngredientsStyles.price} mt-2 mb-2`}>
           <span className="text text_type_digits-default">{props.data.price}</span>
-          <CurrencyIcons  icon={<CurrencyIcon type="primary" />}  alt="Валюта" />
+          <span className={IngredientsStyles.currency}><CurrencyIcon /></span>
         </div>
         <p className="text text text_type_main-default">{props.data.name}</p>
       </li>
