+++ src/components/burger-ingredients/ingredient/ingredient.module.css
@@ -17,5 +17,6 @@
     width: 24px;
     height: 24px;
     
-    transform: translate(-45%,0%);
-}
+    transform: translate(8px,0%);
+} 
+
